apps/qd_plat/requirements.txt
SQLAlchemy~=1.4.29
PyMySQL~=0.9.3
requests~=2.26.0
loguru~=0.6.0
pycryptodome~=3.14.1
pandas~=2.0.0
SQLAlchemy-Utils==0.38.2
python-multipart==0.0.5
schedule~=1.1.0